consolidate duplicated launcher settings logic into dolphin-config
getLauncherDir, readLauncherSettings, and detectDolphinVariant were
independently defined in both dolphin-config.ts and dolphin-launcher.ts.
Move the canonical versions to dolphin-config and import them in
dolphin-launcher, eliminating three copies of the same logic.

Made-with: Cursor
